gingacc-system: Separate unlock() from copyToSD()
This are logically distinct operations and there is
no reason to unify them.
Plus, the copyToSD() is a heavy operation and should
not be hidden, but explicitly called.

With this patch there could be places left where
the copy is no longer done, although it's needed,
so this should be tested carefully.

Signed-off-by: Ezequiel Garcia <elezegarcia@gmail.com>